# openapi.yaml
openapi: 3.1.0
info:
  title: Bijbels-Pastoraat-NL Backend
  version: "1.0.0"
  description: |
    API voor berijmde psalmverzen (1773) via psalmboek.nl.
    Endpoints:
      - GET /api/psalm/vers?psalm={1..150}&vers={1..}
      - GET /api/psalm/max?psalm={1..150}
servers:
  - url: https://gpt-harbers.duckdns.org

paths:
  /api/psalm/vers:
    get:
      summary: Haal exact één vers op uit berijming 1773
      operationId: get_berijmd_psalmvers
      parameters:
        - in: query
          name: psalm
          required: true
          schema: { type: integer, minimum: 1, maximum: 150 }
          description: Psalmnummer 1..150
        - in: query
          name: vers
          required: true
          schema: { type: integer, minimum: 1 }
          description: Versnummer 1..max
      responses:
        "200":
          description: Succes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PsalmVersResponse"
        "400": { description: Buiten bereik of niet gevonden }
        "404": { description: Niet gevonden }
        "422": { description: Validatiefout (query-parameters onjuist) }
        "502": { description: Fout bij bron }

  /api/psalm/max:
    get:
      summary: Bepaal het maximale versnummer van een psalm (1773)
      operationId: get_psalm_max
      parameters:
        - in: query
          name: psalm
          required: true
          schema: { type: integer, minimum: 1, maximum: 150 }
          description: Psalmnummer 1..150
      responses:
        "200":
          description: Succes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PsalmMaxResponse"
        "422": { description: Validatiefout (query-parameters onjuist) }
        "502": { description: Fout bij bron }

components:
  schemas:
    PsalmVersResponse:
      type: object
      properties:
        psalm: { type: integer }
        vers:  { type: integer }
        text:  { type: string }
        bron:  { type: string, format: uri }
      required: [psalm, vers, text, bron]

    PsalmMaxResponse:
      type: object
      properties:
        psalm:    { type: integer }
        max_vers: { type: integer }
        bron:     { type: string, format: uri }
      required: [psalm, max_vers, bron]